pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "cinereserve"
version = "0.1.0"
description = "CineReserve API - Cinema reservation system"
authors = [
    {name = "Andre Barbosa"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "python-decouple (>=3.8,<4.0)",
    "gunicorn (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "coverage (>=7.13.5,<8.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "django-stubs (>=6.0.1,<7.0.0)"
]
